@@ -1,14 +1,8 @@
-import 'package:json_annotation/json_annotation.dart';
-
-import 'response.dart';
-part 'language.g.dart';
-
 /// Encapsulates the list of supported what3words languages
-@JsonSerializable(explicitToJson: true)
-class Language extends Response<Language> {
+class Language {
   List<Languages> languages;
 
-  Language({this.languages});
+  Language({required this.languages});
 
   factory Language.fromJson(Map<String, dynamic> json) =>
       _$LanguageFromJson(json);
@@ -17,16 +11,43 @@ class Language extends Response<Language> {
 }
 
 /// A specific language supported by what3words
-@JsonSerializable(explicitToJson: true)
 class Languages {
   String nativeName;
   String code;
   String name;
 
-  Languages({this.nativeName, this.code, this.name});
+  Languages({required this.nativeName, required this.code, required this.name});
 
   factory Languages.fromJson(Map<String, dynamic> json) =>
       _$LanguagesFromJson(json);
 
   Map<String, dynamic> toJson() => _$LanguagesToJson(this);
 }
+
+//toJson, fromJson
+Language _$LanguageFromJson(Map<String, dynamic> json) {
+  return Language(
+    languages: (json['languages'] as List<dynamic>)
+        .map((e) => Languages.fromJson(e as Map<String, dynamic>))
+        .toList(),
+  );
+}
+
+Map<String, dynamic> _$LanguageToJson(Language instance) => <String, dynamic>{
+  'languages': instance.languages.map((e) => e.toJson()).toList(),
+};
+
+Languages _$LanguagesFromJson(Map<String, dynamic> json) {
+  return Languages(
+    nativeName: json['nativeName'] as String,
+    code: json['code'] as String,
+    name: json['name'] as String,
+  );
+}
+
+Map<String, dynamic> _$LanguagesToJson(Languages instance) => <String, dynamic>{
+  'nativeName': instance.nativeName,
+  'code': instance.code,
+  'name': instance.name,
+};
+
