@@ -1,12 +1,7 @@
-import 'package:json_annotation/json_annotation.dart';
-
 import 'coordinates.dart';
-import 'response.dart';
-part 'location.g.dart';
 
 /// Encapsulates a what3words location
-@JsonSerializable(explicitToJson: true)
-class Location extends Response<Location> {
+class Location {
   String country;
   Square square;
   String nearestPlace;
@@ -16,13 +11,13 @@ class Location extends Response<Location> {
   String map;
 
   Location(
-      {this.country,
-      this.square,
-      this.nearestPlace,
-      this.coordinates,
-      this.words,
-      this.language,
-      this.map});
+      {required this.country,
+      required this.square,
+      required this.nearestPlace,
+      required this.coordinates,
+      required this.words,
+      required this.language,
+      required this.map});
 
   factory Location.fromJson(Map<String, dynamic> json) =>
       _$LocationFromJson(json);
@@ -31,14 +26,52 @@ class Location extends Response<Location> {
 }
 
 /// A what3words square
-@JsonSerializable(explicitToJson: true)
 class Square {
   CoordinatesResponse southwest;
   CoordinatesResponse northeast;
 
-  Square({this.southwest, this.northeast});
+  Square({required this.southwest, required this.northeast});
 
   factory Square.fromJson(Map<String, dynamic> json) => _$SquareFromJson(json);
 
   Map<String, dynamic> toJson() => _$SquareToJson(this);
 }
+
+//toJson, fromJson
+
+Location _$LocationFromJson(Map<String, dynamic> json) {
+  return Location(
+    country: json['country'] as String,
+    square: Square.fromJson(json['square'] as Map<String, dynamic>),
+    nearestPlace: json['nearestPlace'] as String,
+    coordinates: CoordinatesResponse.fromJson(
+        json['coordinates'] as Map<String, dynamic>),
+    words: json['words'] as String,
+    language: json['language'] as String,
+    map: json['map'] as String,
+  );
+}
+
+Map<String, dynamic> _$LocationToJson(Location instance) => <String, dynamic>{
+  'country': instance.country,
+  'square': instance.square.toJson(),
+  'nearestPlace': instance.nearestPlace,
+  'coordinates': instance.coordinates.toJson(),
+  'words': instance.words,
+  'language': instance.language,
+  'map': instance.map,
+};
+
+Square _$SquareFromJson(Map<String, dynamic> json) {
+  return Square(
+    southwest:
+    CoordinatesResponse.fromJson(json['southwest'] as Map<String, dynamic>),
+    northeast:
+    CoordinatesResponse.fromJson(json['northeast'] as Map<String, dynamic>),
+  );
+}
+
+Map<String, dynamic> _$SquareToJson(Square instance) => <String, dynamic>{
+  'southwest': instance.southwest.toJson(),
+  'northeast': instance.northeast.toJson(),
+};
